+++ chk/modules/http/constants.py
@@ -1,0 +1,45 @@
+"""
+Constants used in http modules
+"""
+from enum import Enum
+
+
+class HttpMethod(Enum):
+    """Constants of wellknown http methods"""
+    GET = 'GET'
+    POST = 'POST'
+    PUT = 'PUT'
+    PATCH = 'PATCH'
+    DELETE = 'DELETE'
+    HEAD = 'HEAD'
+    OPTIONS = 'OPTIONS'
+
+
+class BaseDocElements:
+    """represent the base of all kind of documents"""
+    VERSION = 'version'
+
+
+class HttpDocElements(BaseDocElements):
+    """represent http documents"""
+    # common request
+    URL = 'url'
+    METHOD = 'method'
+    HEADERS = 'headers'
+    PARAMS = 'url_params'
+
+    # Basic
+    AUTH_BA = 'auth[basic]'
+    AUTH_BA_USR = 'username'
+    AUTH_BA_PAS = 'password'
+
+    # Bearer
+    AUTH_BE = 'auth[bearer]'
+    AUTH_BE_TOK = 'token'
+
+    # Body
+    BODY_NO = 'body[none]'
+    BODY_FRM = 'body[form]'
+    BODY_FRM_DAT = 'body[form-data]'
+    BODY_JSN = 'body[json]'
+    BODY_XML = 'body[xml]'
